Refactor Client to act as a proxy over API
Instead of having duplicated methods in Client and API
just let the Client proxy via `__call` when the signature is equal
and the body is just forwarding the call anyway.

This makes it easier to add methods and to keep method signatures
in sync. It also puts emphasis on methods that need extra processing
in the Client before hitting the API